requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
pdf2image==1.16.3
Pillow==9.5.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
urllib3==2.0.2
whitenoise==5.3.0
